refactor(server): rename session persistence types (drop "SessionMetadata")
Free up the plain `SessionRepository` name for the server's per-session
store and drop the "SessionMetadata" naming entirely, mirroring
Project/ProjectRepository:

- SessionMetadataRepository -> SessionRepository (file metadata.ts ->
  repository.ts, test renamed to match)
- SessionMetadata (DTO) -> Session
- SessionMetadataNotFound (error + tag) -> SessionNotFound

The harness per-backend cold history store, when it lands, takes the
`HarnessSessionRepository` name.
